# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_geraldthewes_python-executor_pkg_client.AsyncResponse:
    properties:
      execution_id:
        type: string
    type: object
  github_com_geraldthewes_python-executor_pkg_client.CodeFile:
    properties:
      content:
        description: file content
        type: string
      name:
        description: filename (e.g., "main.py")
        type: string
    type: object
  github_com_geraldthewes_python-executor_pkg_client.ExecutionConfig:
    properties:
      cpu_shares:
        type: integer
      disk_mb:
        type: integer
      memory_mb:
        type: integer
      network_disabled:
        type: boolean
      timeout_seconds:
        type: integer
    type: object
  github_com_geraldthewes_python-executor_pkg_client.ExecutionResult:
    properties:
      duration_ms:
        type: integer
      error:
        type: string
      execution_id:
        type: string
      exit_code:
        type: integer
      finished_at:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_geraldthewes_python-executor_pkg_client.ExecutionStatus'
      stderr:
        type: string
      stdout:
        type: string
    type: object
  github_com_geraldthewes_python-executor_pkg_client.ExecutionStatus:
    enum:
    - pending
    - running
    - completed
    - failed
    - killed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusRunning
    - StatusCompleted
    - StatusFailed
    - StatusKilled
  github_com_geraldthewes_python-executor_pkg_client.KillResponse:
    properties:
      status:
        type: string
    type: object
  github_com_geraldthewes_python-executor_pkg_client.SimpleExecRequest:
    properties:
      code:
        description: |-
          Code is the source code to execute (for single-file execution)
          If provided, creates a main.py with this content
        type: string
      config:
        allOf:
        - $ref: '#/definitions/github_com_geraldthewes_python-executor_pkg_client.ExecutionConfig'
        description: Config contains execution resource limits
      entrypoint:
        description: Entrypoint is the file to execute (defaults to "main.py")
        type: string
      files:
        description: |-
          Files allows multiple files to be provided (Piston-compatible)
          Takes precedence over Code if both are provided
        items:
          $ref: '#/definitions/github_com_geraldthewes_python-executor_pkg_client.CodeFile'
        type: array
      stdin:
        description: Stdin is the standard input to provide to the script
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: python-executor
    url: https://github.com/geraldthewes/python-executor
  description: |-
    Remote Python code execution service.

    IMPORTANT: Use the client libraries instead of implementing this API directly.
    The API uses multipart/form-data with tar archives, which is complex to implement correctly.

    Python: pip install git+https://github.com/geraldthewes/python-executor.git#subdirectory=python
    Go: go get github.com/geraldthewes/python-executor/pkg/client
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: python-executor API
  version: "1.0"
paths:
  /eval:
    post:
      consumes:
      - application/json
      description: |-
        Execute Python code using a simple JSON interface.
        This endpoint is designed for AI agents and simple integrations.

        Two modes are supported:
        - Single file: provide "code" field with Python code
        - Multi-file: provide "files" array with name/content pairs
      parameters:
      - description: Execution request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_geraldthewes_python-executor_pkg_client.SimpleExecRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Execution completed
          schema:
            $ref: '#/definitions/github_com_geraldthewes_python-executor_pkg_client.ExecutionResult'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/gin.H'
        "413":
          description: Code size exceeds limit
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Execution failed
          schema:
            $ref: '#/definitions/gin.H'
      summary: Execute code via JSON (simplified API)
      tags:
      - execution
  /exec/async:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Submit code for execution and return immediately with an execution ID.

        IMPORTANT: Use the client libraries instead of calling this directly.
        The request must be multipart/form-data with a tar archive and metadata JSON.
      parameters:
      - description: Uncompressed tar archive containing Python files
        in: formData
        name: tar
        required: true
        type: file
      - description: 'Execution metadata as JSON: {\'
        in: formData
        name: metadata
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Execution submitted
          schema:
            $ref: '#/definitions/github_com_geraldthewes_python-executor_pkg_client.AsyncResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Failed to create execution
          schema:
            $ref: '#/definitions/gin.H'
      summary: Execute code asynchronously
      tags:
      - execution
  /exec/sync:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Execute Python code and wait for result.

        IMPORTANT: Use the client libraries instead of calling this directly.
        The request must be multipart/form-data with a tar archive and metadata JSON.
      parameters:
      - description: Uncompressed tar archive containing Python files
        in: formData
        name: tar
        required: true
        type: file
      - description: 'Execution metadata as JSON: {\'
        in: formData
        name: metadata
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Execution completed
          schema:
            $ref: '#/definitions/github_com_geraldthewes_python-executor_pkg_client.ExecutionResult'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Execution failed
          schema:
            $ref: '#/definitions/gin.H'
      summary: Execute code synchronously
      tags:
      - execution
  /executions/{id}:
    delete:
      description: |-
        Terminate a running execution.
        If the execution is not running, returns the current status.
      parameters:
      - description: Execution ID (e.g., exe_550e8400-e29b-41d4-a716-446655440000)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Execution killed or current status
          schema:
            $ref: '#/definitions/github_com_geraldthewes_python-executor_pkg_client.KillResponse'
        "404":
          description: Execution not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Failed to kill execution
          schema:
            $ref: '#/definitions/gin.H'
      summary: Kill execution
      tags:
      - execution
    get:
      description: |-
        Retrieve the status and result of an execution.
        Status values: pending, running, completed, failed, killed
      parameters:
      - description: Execution ID (e.g., exe_550e8400-e29b-41d4-a716-446655440000)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Execution status and result
          schema:
            $ref: '#/definitions/github_com_geraldthewes_python-executor_pkg_client.ExecutionResult'
        "404":
          description: Execution not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get execution status
      tags:
      - execution
swagger: "2.0"